+++ tests/src/Project/Console/IntegrationTestCase.php
@@ -2,13 +2,13 @@
 namespace Project\Console;
 
 use Opulence\Bootstrappers\ApplicationBinder;
-use Opulence\Framework\Testing\PhpUnit\Console\ApplicationTestCase as BaseTestCase;
+use Opulence\Framework\Testing\PhpUnit\Console\IntegrationTestCase as BaseIntegrationTestCase;
 use Opulence\Ioc\IContainer;
 
 /**
- * Defines the console application test case
+ * Defines the console application integration test
  */
-class ApplicationTestCase extends BaseTestCase
+class IntegrationTestCase extends BaseIntegrationTestCase
 {
     /**
      * @inheritdoc